+++ pkg/gds/pagination/v1/generate.go
@@ -1,0 +1,3 @@
+package pagination
+
+//go:generate protoc -I=$GOPATH/src/github.com/trisacrypto/directory/proto --go_out=. --go_opt=module=github.com/trisacrypto/directory/pkg/gds/pagination/v1  gds/pagination/v1/pagination.proto
