+++ pkg/gds/pagination/v1/pagination.go
@@ -1,4 +1,4 @@
-package models
+package pagination
 
 import (
 	"encoding/base64"
