+++ proto/gds/models/v1/models.proto
@@ -73,14 +73,3 @@ message ReviewNote {
     // Actual text in the note
     string text = 6;
 }
-
-// Implements a protocol buffer struct for state managed pagination. This struct will be
-// marshaled into a url-safe base64 encoded string and sent to the user as the
-// next_page_token. The server should decode this struct to determine where to continue
-// iteration for the next page. Note that the server should check to make sure the page
-// size in the cursor matches the page size in the request.
-// See https://cloud.google.com/apis/design/design_patterns#list_pagination for more.
-message PageCursor {
-    int32 page_size = 1;  // the number of results returned on each iteration.
-    string next_vasp = 2; // the VASP id to start the iteration from
-}
